add api foudations and update swagger

+++ src/internal/sql_scripts/create.sql
@@ -7,8 +7,8 @@ CREATE TABLE USER_TAB(
                      id integer NOT NULL PRIMARY KEY,
                     login varchar(100)  NOT NULL,
 					password varchar(100)  NOT NULL,
-					balance money,
-                    charity_Sum money ,
+					balance REAL,
+                    charity_Sum REAL ,
                     closed_Fing_Amount integer
 					);
 
@@ -18,9 +18,9 @@ CREATE TABLE foundation_TAB(
 					password varchar(100)  NOT NULL,
                     cur_Foudrising_Amount integer,
                     closed_Foundrising_Amount integer,
-                    fund_balance money,
-                    income_history money,
-                    outcome_history money,
+                    fund_balance REAL,
+                    income_history REAL,
+                    outcome_history REAL,
                     volunteer_amount integer,
 					country varchar(100),
 					login varchar(100)  NOT NULL);
@@ -30,8 +30,8 @@ CREATE TABLE foundrising_TAB(
                     id integer NOT NULL PRIMARY KEY,
                     found_id integer NOT NULL,
                     description text,	
-                    required_sum money,
-                    current_sum money,
+                    required_sum REAL,
+                    current_sum REAL,
                     philantrops_amount integer,
                     creation_date date,
 					closing_date date);
@@ -41,7 +41,7 @@ CREATE TABLE transaction_TAB(
 					from_essence_type boolean, -- 0 - user 1 - foundation
                     from_id integer NOT NULL,
                     to_essence_type boolean, -- 0 - foundation, 1 - foundrising
-                    sum_of_money money,
+                    sum_of_money REAL,
                     comment text,
                     to_id integer NOT NULL);
